mabe-solution/requirements.txt
# Core PyTorch
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Lightning and training
pytorch-lightning>=2.0.0
torchmetrics>=0.11.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
pyarrow>=10.0.0
scipy>=1.9.0
scikit-learn>=1.2.0

# Configuration and experiment tracking
hydra-core>=1.3.0
wandb>=0.15.0

# Utilities
tqdm>=4.64.0
matplotlib>=3.6.0
seaborn>=0.12.0
joblib>=1.2.0

# Colab specific (optional)
notebook>=6.5.0
